perf(rasterizer): optimize memory usage and computation with half precision
- Change uint to ushort for contribution counters to reduce memory footprint
- Add max_contributions limit to prevent overflow
- Implement half precision optimizations in forward and backward passes
- Simplify PerPixel struct layout for better memory efficiency
- Update config to use float32 for dataloader and trainer

+++ tinygs/src/rasterizer/fastgs_ours_fp16/buffer_utils.h
@@ -112,8 +112,8 @@ struct PerTileBuffers {
   uint2* instance_ranges;
   uint* n_buckets;
   uint* bucket_offsets;
-  uint* max_n_contributions;
-  uint* n_contributions;
+  ushort* max_n_contributions;
+  ushort* n_contributions;
 
   static PerTileBuffers from_blob(char*& blob, size_t n_tiles) {
     PerTileBuffers buffers;
